requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coverage==4.5.2
cryptography==2.4.2
Django==2.1.11
django-cors-headers==2.4.0
idna==2.7
Pillow==5.3.0
pycparser==2.19
PyJWT==1.7.1
pytz==2018.5
requests==2.19.1
six==1.12.0
urllib3==1.23
